detekt lint issues fixed.
