+++ core/src/main/java/dev/happysingh/core/ext/CommonViewExt.kt
@@ -1,0 +1,58 @@
+package dev.happysingh.core.ext
+
+import android.view.View
+import android.view.ViewGroup
+import com.google.android.material.snackbar.Snackbar
+
+operator fun ViewGroup.get(index: Int): View {
+    return getChildAt(index)
+}
+
+fun View.visible() {
+    if (visibility != View.VISIBLE) {
+        visibility = View.VISIBLE
+    }
+}
+
+inline fun View.visibleIf(block: () -> Boolean) {
+    if (visibility != View.VISIBLE && block()) {
+        visibility = View.VISIBLE
+    }
+}
+
+fun View.invisible() {
+    if (visibility != View.INVISIBLE) {
+        visibility = View.INVISIBLE
+    }
+}
+
+inline fun View.invisiableIf(block: () -> Boolean) {
+    if (visibility != View.INVISIBLE && block()) {
+        visibility = View.INVISIBLE
+    }
+}
+
+fun View.gone() {
+    if (visibility != View.GONE) {
+        visibility = View.GONE
+    }
+}
+
+inline fun View.goneIf(block: () -> Boolean) {
+    if (visibility != View.GONE && block()) {
+        visibility = View.GONE
+    }
+}
+
+fun View.isVisible() = visibility == View.VISIBLE
+
+fun View.isGone() = visibility == View.GONE
+
+fun View.isInvisible() = visibility == View.INVISIBLE
+
+/**
+ * This showSnackBar fun can be called from any view
+ */
+fun View.showSnackBar(message: String) {
+    Snackbar.make(this, message, Snackbar.LENGTH_LONG).show()
+}
