--- core/src/main/java/dev/happysingh/core/ext/ViewExt.kt
@@ -1,283 +1,0 @@
-package dev.happysingh.core.ext
-
-import android.content.Context
-import android.graphics.Bitmap
-import android.graphics.Canvas
-import android.graphics.Paint
-import android.text.method.PasswordTransformationMethod
-import android.text.method.ReplacementTransformationMethod
-import android.view.View
-import android.view.ViewGroup
-import android.view.animation.AccelerateDecelerateInterpolator
-import android.view.animation.Interpolator
-import android.webkit.WebChromeClient
-import android.webkit.WebResourceError
-import android.webkit.WebResourceRequest
-import android.webkit.WebView
-import android.webkit.WebViewClient
-import android.widget.EditText
-import android.widget.RadioButton
-import android.widget.RadioGroup
-import android.widget.ScrollView
-import android.widget.TextView
-import androidx.annotation.ColorRes
-import androidx.appcompat.widget.AppCompatButton
-import androidx.appcompat.widget.AppCompatEditText
-import androidx.appcompat.widget.AppCompatTextView
-import androidx.core.content.ContextCompat
-import androidx.core.graphics.ColorUtils
-import androidx.core.widget.NestedScrollView
-import kotlin.math.roundToInt
-
-private const val duration = 750L
-private inline val interpolator: Interpolator
-    get() = AccelerateDecelerateInterpolator()
-
-fun View.setPaddingLeft(value: Int) = setPadding(value, paddingTop, paddingRight, paddingBottom)
-
-fun View.setPaddingRight(value: Int) = setPadding(paddingLeft, paddingTop, value, paddingBottom)
-
-fun View.setPaddingTop(value: Int) = setPaddingRelative(paddingStart, value, paddingEnd, paddingBottom)
-
-fun View.setPaddingBottom(value: Int) = setPaddingRelative(paddingStart, paddingTop, paddingEnd, value)
-
-fun View.setPaddingStart(value: Int) = setPaddingRelative(value, paddingTop, paddingEnd, paddingBottom)
-
-fun View.setPaddingEnd(value: Int) = setPaddingRelative(paddingStart, paddingTop, value, paddingBottom)
-
-fun View.setPaddingHorizontal(value: Int) = setPaddingRelative(value, paddingTop, value, paddingBottom)
-
-fun View.setPaddingVertical(value: Int) = setPaddingRelative(paddingStart, value, paddingEnd, value)
-
-fun View.setHeight(value: Int) {
-    val lp = layoutParams
-    lp?.let {
-        lp.height = value
-        layoutParams = lp
-    }
-}
-
-fun View.setWidth(value: Int) {
-    val lp = layoutParams
-    lp?.let {
-        lp.width = value
-        layoutParams = lp
-    }
-}
-
-fun View.resize(width: Int, height: Int) {
-    val lp = layoutParams
-    lp?.let {
-        lp.width = width
-        lp.height = height
-        layoutParams = lp
-    }
-}
-
-inline val ViewGroup.children: List<View>
-    get() = (0 until childCount).map { getChildAt(it) }
-
-operator fun ViewGroup.get(index: Int): View {
-    return getChildAt(index)
-}
-
-fun TextView.underLine() {
-    paint.flags = paint.flags or Paint.UNDERLINE_TEXT_FLAG
-    paint.isAntiAlias = true
-}
-
-fun TextView.deleteLine() {
-    paint.flags = paint.flags or Paint.STRIKE_THRU_TEXT_FLAG
-    paint.isAntiAlias = true
-}
-
-fun TextView.bold(isBold: Boolean = true) {
-    paint.isFakeBoldText = isBold
-    paint.isAntiAlias = true
-}
-
-fun <T : View> T.click(block: (T) -> Unit) = setOnClickListener { block(it as T) }
-
-fun <T : View> T.longClick(block: (T) -> Boolean) = setOnLongClickListener { block(it as T) }
-
-fun View.visible() {
-    if (visibility != View.VISIBLE) {
-        visibility = View.VISIBLE
-    }
-}
-
-inline fun View.visibleIf(block: () -> Boolean) {
-    if (visibility != View.VISIBLE && block()) {
-        visibility = View.VISIBLE
-    }
-}
-
-fun View.invisible() {
-    if (visibility != View.INVISIBLE) {
-        visibility = View.INVISIBLE
-    }
-}
-
-inline fun View.invisiableIf(block: () -> Boolean) {
-    if (visibility != View.INVISIBLE && block()) {
-        visibility = View.INVISIBLE
-    }
-}
-
-fun View.gone() {
-    if (visibility != View.GONE) {
-        visibility = View.GONE
-    }
-}
-
-inline fun View.goneIf(block: () -> Boolean) {
-    if (visibility != View.GONE && block()) {
-        visibility = View.GONE
-    }
-}
-
-var EditText.value: String
-    get() = text.toString()
-    set(value) = setText(value)
-
-fun View.getBitmap(): Bitmap {
-    val bmp = Bitmap.createBitmap(width, height, Bitmap.Config.ARGB_8888)
-    val canvas = Canvas(bmp)
-    draw(canvas)
-    canvas.save()
-    return bmp
-}
-
-fun EditText.uppercase() {
-    transformationMethod = object : ReplacementTransformationMethod() {
-        private val lower = charArrayOf('a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z')
-        private val upper = charArrayOf('A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z')
-
-        override fun getOriginal() = lower
-
-        override fun getReplacement() = upper
-    }
-}
-
-fun EditText.lowercase() {
-    transformationMethod = object : ReplacementTransformationMethod() {
-        private val lower = charArrayOf('a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z')
-        private val upper = charArrayOf('A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z')
-
-        override fun getOriginal() = upper
-
-        override fun getReplacement() = lower
-    }
-}
-
-fun EditText.passwordToggledVisible() {
-    val selection = selectionStart
-    transformationMethod = if (transformationMethod == null) PasswordTransformationMethod() else null
-    setSelection(selection)
-}
-
-fun View.isVisible() = visibility == View.VISIBLE
-
-fun View.isGone() = visibility == View.GONE
-
-fun View.isInvisible() = visibility == View.INVISIBLE
-
-var RadioGroup.checkedIndex: Int
-    get() = (0 until childCount).firstOrNull { (getChildAt(it) as RadioButton).isChecked } ?: -1
-    set(value) {
-        if (value !in 0 until childCount) {
-            children.map { it as RadioButton }.filter { it.isChecked }.forEach { it.isChecked = false }
-        } else {
-            (getChildAt(value) as RadioButton).isChecked = true
-        }
-    }
-
-fun AppCompatButton.disableAlpha() {
-    this.isClickable = false
-    this.isEnabled = false
-    this.alpha = 0.5f
-}
-
-fun AppCompatButton.enableWithDefaultAlpha() {
-    this.isClickable = true
-    this.isEnabled = true
-    this.alpha = 1f
-}
-
-fun AppCompatTextView.disableAlpha() {
-    this.isClickable = false
-    this.isEnabled = false
-    this.alpha = 0.5f
-}
-
-fun AppCompatTextView.enableWithDefaultAlpha() {
-    this.isClickable = true
-    this.isEnabled = true
-    this.alpha = 1f
-}
-
-fun AppCompatEditText.disableAlpha() {
-    alpha = 0.5f
-    isFocusable = false
-    isFocusableInTouchMode = false
-}
-
-fun AppCompatEditText.enableWithDefaultAlpha() {
-    alpha = 1f
-    isFocusable = true
-    isFocusableInTouchMode = true
-}
-
-/**
- * returns lambda if [WebView] is successfully loaded.
- */
-fun WebView.doOnPageLoad(onWebPageLoad: () -> Unit) {
-    webChromeClient = object : WebChromeClient() {
-
-        override fun onProgressChanged(view: WebView?, newProgress: Int) {
-            super.onProgressChanged(view, newProgress)
-            if (progress == 100) {
-                onWebPageLoad.invoke()
-            }
-        }
-    }
-}
-
-/**
- * returns lambda if [WebView] is gets an error while loading loaded.
- */
-fun WebView.doOnPageLoadError(onError: () -> Unit) {
-    webViewClient = object : WebViewClient() {
-
-        override fun onReceivedError(
-            view: WebView?,
-            request: WebResourceRequest?,
-            error: WebResourceError?
-        ) {
-            super.onReceivedError(view, request, error)
-            onError.invoke()
-        }
-    }
-}
-
-fun ScrollView.scrollToBottom() {
-    val lastChild = getChildAt(childCount - 1)
-    val bottom = lastChild.bottom + paddingBottom
-    val delta = bottom - (scrollY + height)
-    smoothScrollBy(0, delta)
-}
-
-fun NestedScrollView.scrollToBottom() {
-    val lastChild = getChildAt(childCount - 1)
-    val bottom = lastChild.bottom + paddingBottom
-    val delta = bottom - (scrollY + height)
-    smoothScrollBy(0, delta)
-}
-
-// send alpha under from 0.0f to 1.0f.
-fun Context.getAlphaColor(@ColorRes color: Int, alpha: Float): Int {
-    return ColorUtils.setAlphaComponent(
-        ContextCompat.getColor(this, color),
-        255.times(alpha).roundToInt()
-    )
-}
